requirements.txt
pandas
numpy
biopython
matplotlib